requirements.txt
aiogram==3.*
aiohttp
aiocron
databases[aiomysql]
python-dotenv
